refactor: move vllm_sidecar package to repo top level.
