+++ vllm_sidecar/etcd_registry.py
